Feat/Refactor: More conditions for detecting end
Stalemate now declared if player only has king, which cannot move,
but is not threatened either.

Enforce check now takes input to check from both player PoVs

Game winner now takes input rather than turn based choice

Improved king safety check, to be safe the following must be true:
1) The king has a valid move outside of the "urgent" PoV squares.
A move is not valid if that move ALSO would be in PoV of another piece.

AND that move is not taking an enemy piece that a enemy can see.
(previously a move that involves taking an enemy views as safe, because
the enemy technically did not include their ally in their valid moves)

The method for handling this is very explicit and seperated out,
for a proof of concept. This should be refactored and split into more objects,
alongside other similar methods that do almost the same.