Refactor(Misc): small refactors, see body for info
- Validate king safety after move now passes a block to DRY.
Could be used to track any piece in particular.
Will experiment with using this during checkmate-validation,
to reduce queries by searching for specific pieces.

- Can no longer castle if game is in check.  No need to validate who is
in check, because during a check only the defending player
can be in control.

-Attack Moveset is now a valid piece method.
In most cases is an alias for moveset.
However kings and pawns have a method to override.

Cannot be an alias in piece.rb due to the split class inheritence

- Removed methods previously made obsolete in previous commits

- Game view properly displays a tie