Reactor & Bugfix: King safety checking. See body!
Previously, check(mate) was  marked as true if a piece moved
 into a position that allowed them to see the king.
This meant that checking a king by moving a piece
which gives another friendly piece LoS would not mark check.
This also meant that if it was checkmate, you would softlock
the opponent, since the game would deny them any unsafe turns
but did not realize there were 0 safe turns.

Now:

Wrote a method that inspects each potential move the king could make
(King.rb::king_is_in_sights)
Returns moves that an enemy could attack,using the respective moveset.

The GUI for king move select, the evaluation for if check has begun,
the evaluation for if game is over, and the evalution for rollback
can all plug into this method to see if the king is safe at any point.

This is significantly faster than the previous solution of
querying every enemy piece for if they could spot the king or not
when filtering out king moves or checking safety for rollback.

Movesets have also been factored back out to Piece.rb so
that this evaluation can use them without needing to repeat them.